Fix failing tests
* Fix failing feature flag integration tests.
