+++ src/main/java/uk/gov/companieshouse/certificates/orders/api/controller/ApiErrors.java
@@ -95,6 +95,17 @@ public final class ApiErrors {
     private static final String SECRETARY_DETAILS_INCLUDE_OCCUPATION_LOCATION = "secretary_details.include_occupation";
     private static final String SECRETARY_DETAILS_INCLUDE_DOB_TYPE_LOCATION = "secretary_details.include_dob_type";
 
+    private static final String DESIGNATED_MEMBER_DETAILS_INCLUDE_ADDRESS_LOCATION = "designated_member_details.include_address";
+    private static final String DESIGNATED_MEMBER_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION = "designated_member_details.include_appointment_date";
+    private static final String DESIGNATED_MEMBER_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION = "designated_member_details.include_country_of_residence";
+    private static final String DESIGNATED_MEMBER_DETAILS_INCLUDE_DOB_TYPE_LOCATION = "designated_member_details.include_dob_type";
+
+    private static final String MEMBER_DETAILS_INCLUDE_ADDRESS_LOCATION = "member_details.include_address";
+    private static final String MEMBER_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION = "member_details.include_appointment_date";
+    private static final String MEMBER_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION = "member_details.include_country_of_residence";
+    private static final String MEMBER_DETAILS_INCLUDE_DOB_TYPE_LOCATION = "member_details.include_dob_type";
+
+
     public static final String STRING_LOCATION_TYPE = "string";
     public static final String BOOLEAN_LOCATION_TYPE = "boolean";
     public static final String OBJECT_LOCATION_TYPE = "object";
@@ -136,18 +147,33 @@ public final class ApiErrors {
     public static final ApiError ERR_QUANTITY_AMOUNT = new ApiError(QUANTITY_AMOUNT_ERROR, QUANTITY_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DESCRIPTION_IDENTIFIER = new ApiError(DESCRIPTION_IDENTIFIER_ERROR, DESCRIPTION_IDENTIFIER_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_POSTAL_DELIVERY = new ApiError(POSTAL_DELIVERY_ERROR, POSTAL_DELIVERY_LOCATION, OBJECT_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_ADDRESS = new ApiError(INCLUDE_ADDRESS_ERROR, DIRECTOR_DETAILS_INCLUDE_ADDRESS_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_APPOINTMENT_DATE = new ApiError(INCLUDE_APPOINTMENT_DATE_ERROR, DIRECTOR_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_ERROR = new ApiError(INCLUDE_COUNTRY_OF_RESIDENCE_ERROR, DIRECTOR_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_NATIONALITY = new ApiError(INCLUDE_NATIONALITY_ERROR, DIRECTOR_DETAILS_INCLUDE_NATIONALITY_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_OCCUPATION = new ApiError(INCLUDE_OCCUPATION_ERROR, DIRECTOR_DETAILS_INCLUDE_OCCUPATION_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_DIRECTOR_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_ERROR, DIRECTOR_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_ADDRESS = new ApiError(INCLUDE_ADDRESS_ERROR, SECRETARY_DETAILS_INCLUDE_ADDRESS_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_APPOINTMENT_DATE = new ApiError(INCLUDE_APPOINTMENT_DATE_ERROR, SECRETARY_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_ERROR = new ApiError(INCLUDE_COUNTRY_OF_RESIDENCE_ERROR, SECRETARY_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_NATIONALITY = new ApiError(INCLUDE_NATIONALITY_ERROR, SECRETARY_DETAILS_INCLUDE_NATIONALITY_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_OCCUPATION = new ApiError(INCLUDE_OCCUPATION_ERROR, SECRETARY_DETAILS_INCLUDE_OCCUPATION_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
     public static final ApiError ERR_SECRETARY_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_ERROR, SECRETARY_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+
+    public static final ApiError ERR_DESIGNATED_MEMBER_DETAILS_INCLUDE_ADDRESS = new ApiError(INCLUDE_ADDRESS_ERROR, DESIGNATED_MEMBER_DETAILS_INCLUDE_ADDRESS_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_DESIGNATED_MEMBER_DETAILS_INCLUDE_APPOINTMENT_DATE = new ApiError(INCLUDE_APPOINTMENT_DATE_ERROR, DESIGNATED_MEMBER_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_DESIGNATED_MEMBER_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_ERROR = new ApiError(INCLUDE_COUNTRY_OF_RESIDENCE_ERROR, DESIGNATED_MEMBER_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_DESIGNATED_MEMBER_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_ERROR, DESIGNATED_MEMBER_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+
+    public static final ApiError ERR_MEMBER_DETAILS_INCLUDE_ADDRESS = new ApiError(INCLUDE_ADDRESS_ERROR, MEMBER_DETAILS_INCLUDE_ADDRESS_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_MEMBER_DETAILS_INCLUDE_APPOINTMENT_DATE = new ApiError(INCLUDE_APPOINTMENT_DATE_ERROR, MEMBER_DETAILS_INCLUDE_APPOINTMENT_DATE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_MEMBER_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_ERROR = new ApiError(INCLUDE_COUNTRY_OF_RESIDENCE_ERROR, MEMBER_DETAILS_INCLUDE_COUNTRY_OF_RESIDENCE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+    public static final ApiError ERR_MEMBER_DETAILS_INCLUDE_DOB_TYPE_SUPPLIED_ERROR = new ApiError(INCLUDE_DOB_TYPE_ERROR, MEMBER_DETAILS_INCLUDE_DOB_TYPE_LOCATION, BOOLEAN_LOCATION_TYPE, ERROR_TYPE_VALIDATION);
+
+
+
 
     private ApiErrors() {}
