+++ src/test/java/uk/gov/companieshouse/certificates/orders/api/validator/CreateItemRequestValidatorLPFeatureFlagEnabledIntegrationTest.java
@@ -1,16 +1,28 @@
 package uk.gov.companieshouse.certificates.orders.api.validator;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.empty;
+import static org.hamcrest.Matchers.is;
+import static org.mockito.Mockito.when;
+import static uk.gov.companieshouse.certificates.orders.api.model.CertificateType.DISSOLUTION;
+import static uk.gov.companieshouse.certificates.orders.api.model.DeliveryMethod.COLLECTION;
+import static uk.gov.companieshouse.certificates.orders.api.model.DeliveryTimescale.SAME_DAY;
+import static uk.gov.companieshouse.certificates.orders.api.model.DeliveryTimescale.STANDARD;
+import static uk.gov.companieshouse.certificates.orders.api.model.IncludeDobType.PARTIAL;
+
+import java.util.Arrays;
+import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
 import uk.gov.companieshouse.api.error.ApiError;
-import uk.gov.companieshouse.certificates.orders.api.config.FeatureOptionsConfig;
 import uk.gov.companieshouse.certificates.orders.api.controller.ApiErrors;
 import uk.gov.companieshouse.certificates.orders.api.model.CertificateItemOptions;
 import uk.gov.companieshouse.certificates.orders.api.model.DesignatedMemberDetails;
@@ -24,38 +36,23 @@ import uk.gov.companieshouse.certificates.orders.api.model.RegisteredOfficeAddre
 import uk.gov.companieshouse.logging.Logger;
 import uk.gov.companieshouse.logging.LoggerFactory;
 
-import java.util.Arrays;
-import java.util.List;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.contains;
-import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.empty;
-import static org.hamcrest.Matchers.is;
-import static org.mockito.Mockito.when;
-import static uk.gov.companieshouse.certificates.orders.api.model.CertificateType.DISSOLUTION;
-import static uk.gov.companieshouse.certificates.orders.api.model.DeliveryMethod.COLLECTION;
-import static uk.gov.companieshouse.certificates.orders.api.model.DeliveryTimescale.SAME_DAY;
-import static uk.gov.companieshouse.certificates.orders.api.model.DeliveryTimescale.STANDARD;
-import static uk.gov.companieshouse.certificates.orders.api.model.IncludeDobType.PARTIAL;
-
 /**
  * Unit tests the {@link CreateItemRequestValidator} class.
  */
-@Import({CertificateOptionsValidatorConfig.class, FeatureOptionsConfig.class})
 @SpringBootTest
-@ActiveProfiles("feature-flags-enabled")
-class CreateItemRequestValidatorFeatureFlagsEnabledTest {
+@ActiveProfiles("lp-feature-flag-enabled")
+class CreateItemRequestValidatorLPFeatureFlagEnabledIntegrationTest {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(CreateItemRequestValidatorFeatureFlagsEnabledTest.class.getName());
-
-    @Mock
-    private RequestValidatable requestValidatable;
-
-    private CertificateItemOptions certificateItemOptions;
+    private static final Logger LOGGER = LoggerFactory.getLogger(CreateItemRequestValidatorLPFeatureFlagEnabledIntegrationTest.class.getName());
 
     @Value("${spring.profiles.active}")
     private String activeProfile;
+
+    private CertificateItemOptions certificateItemOptions;
+
+    @MockBean
+    private RequestValidatable requestValidatable;
+
     @Autowired
     private CreateItemRequestValidator validatorUnderTest;
 
@@ -63,7 +60,6 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     void setUp() {
         LOGGER.debug("Active profile " + activeProfile);
         certificateItemOptions = new CertificateItemOptions();
-
         when(requestValidatable.getItemOptions()).thenReturn(certificateItemOptions);
     }
 
@@ -71,11 +67,9 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     @DisplayName("ID is mandatory")
     void idIsMandatory() {
         // Given
-        certificateItemOptions.setCompanyType("limited");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
+        certificateItemOptions.setCompanyType("limited-partnership");
         when(requestValidatable.getCertificateId()).thenReturn("1");
-        ApiError expectedError = ApiErrors.raiseError(ApiErrors.ERR_CERTIFICATE_ID_SUPPLIED,
-                "id: must be null in a create item request");
+        ApiError expectedError = ApiErrors.raiseError(ApiErrors.ERR_CERTIFICATE_ID_SUPPLIED, "id: must be null in a create item request");
 
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
@@ -88,8 +82,7 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     @DisplayName("Collection location is optional by default")
     void collectionLocationIsOptionalByDefault() {
         // Given
-        certificateItemOptions.setCompanyType("limited");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
+        certificateItemOptions.setCompanyType("limited-partnership");
         certificateItemOptions.setCollectionLocation(null);
 
         // When
@@ -103,7 +96,6 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     @DisplayName("Collection details are mandatory for collection delivery method")
     void collectionDetailsAreMandatoryForCollectionDeliveryMethod() {
         // Given
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
         certificateItemOptions.setDeliveryMethod(COLLECTION);
         certificateItemOptions.setCompanyType("any");
         List<ApiError> expectedErrors = Arrays.asList(
@@ -126,7 +118,7 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         certificateItemOptions.setIncludeCompanyObjectsInformation(true);
         certificateItemOptions.setIncludeGoodStandingInformation(true);
         certificateItemOptions.setCompanyType("any");
-	
+
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
@@ -139,6 +131,10 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
             "secretary details or director details should not be requested for dissolution")
     void certainCompanyObjectsMustNotBeRequestedForDissolution() {
         // Given
+        RegisteredOfficeAddressDetails registeredOfficeAddressDetails = new RegisteredOfficeAddressDetails();
+        registeredOfficeAddressDetails.setIncludeAddressRecordsType(IncludeAddressRecordsType.CURRENT);
+        registeredOfficeAddressDetails.setIncludeDates(true);
+
         DirectorOrSecretaryDetails directorOrSecretaryDetails = new DirectorOrSecretaryDetails();
         directorOrSecretaryDetails.setIncludeAddress(true);
         directorOrSecretaryDetails.setIncludeAppointmentDate(false);
@@ -148,10 +144,12 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         directorOrSecretaryDetails.setIncludeNationality(false);
         directorOrSecretaryDetails.setIncludeOccupation(true);
 
-        RegisteredOfficeAddressDetails registeredOfficeAddressDetails = new RegisteredOfficeAddressDetails();
-        registeredOfficeAddressDetails.setIncludeAddressRecordsType(IncludeAddressRecordsType.CURRENT);
-        registeredOfficeAddressDetails.setIncludeDates(true);
-
+        MemberDetails memberDetails = new MemberDetails();
+        DesignatedMemberDetails designatedMemberDetails = new DesignatedMemberDetails();
+        GeneralPartnerDetails generalPartnerDetails = new GeneralPartnerDetails();
+        LimitedPartnerDetails limitedPartnerDetails = new LimitedPartnerDetails();
+        PrincipalPlaceOfBusinessDetails principalPlaceOfBusinessDetails = new PrincipalPlaceOfBusinessDetails();
+	
         certificateItemOptions.setCertificateType(DISSOLUTION);
         certificateItemOptions.setIncludeCompanyObjectsInformation(true);
         certificateItemOptions.setIncludeGoodStandingInformation(true);
@@ -159,11 +157,11 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         certificateItemOptions.setRegisteredOfficeAddressDetails(registeredOfficeAddressDetails);
         certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
         certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
-        certificateItemOptions.setMemberDetails(new MemberDetails());
-        certificateItemOptions.setDesignatedMemberDetails(new DesignatedMemberDetails());
-        certificateItemOptions.setGeneralPartnerDetails(new GeneralPartnerDetails());
-        certificateItemOptions.setLimitedPartnerDetails(new LimitedPartnerDetails());
-        certificateItemOptions.setPrincipalPlaceOfBusinessDetails(new PrincipalPlaceOfBusinessDetails());
+        certificateItemOptions.setMemberDetails(memberDetails);
+        certificateItemOptions.setDesignatedMemberDetails(designatedMemberDetails);
+        certificateItemOptions.setGeneralPartnerDetails(generalPartnerDetails);
+        certificateItemOptions.setLimitedPartnerDetails(limitedPartnerDetails);
+        certificateItemOptions.setPrincipalPlaceOfBusinessDetails(principalPlaceOfBusinessDetails);
         certificateItemOptions.setCompanyType("limited");
         List<ApiError> expectedErrors = Arrays.asList(
                 ApiErrors.raiseError(ApiErrors.ERR_INCLUDE_COMPANY_OBJECTS_INFORMATION_SUPPLIED, "include_company_objects_information: must not exist when certificate type is dissolution"),
@@ -184,29 +182,47 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
                 ApiErrors.raiseError(ApiErrors.ERR_PRINCIPAL_PLACE_OF_BUSINESS_DETAILS_SUPPLIED, "include_principal_place_of_business_details: must not exist when company type is limited"),
                 ApiErrors.raiseError(ApiErrors.ERR_INCLUDE_GENERAL_NATURE_OF_BUSINESS_INFORMATION_SUPPLIED, "include_general_nature_of_business_information: must not exist when company type is limited")
         );
-	
+
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
         // Then
+//        assertThat(errors, containsInAnyOrder(
+//                "include_company_objects_information: must not exist when certificate type is dissolution",
+//                "include_good_standing_information: must not exist when certificate type is dissolution",
+//                "include_general_nature_of_business_information: must not exist when certificate type is dissolution",
+//                "include_registered_office_address_details: must not exist when certificate type is dissolution",
+//                "include_secretary_details: must not exist when certificate type is dissolution",
+//                "include_director_details: must not exist when certificate type is dissolution",
+//                "include_member_details: must not exist when certificate type is dissolution",
+//                "include_designated_member_details: must not exist when certificate type is dissolution",
+//                "include_general_partner_details: must not exist when certificate type is dissolution",
+//                "include_limited_partner_details: must not exist when certificate type is dissolution",
+//                "include_principal_place_of_business_details: must not exist when certificate type is dissolution",
+//                "include_member_details: must not exist when company type is limited",
+//                "include_designated_member_details: must not exist when company type is limited",
+//                "include_general_partner_details: must not exist when company type is limited",
+//                "include_limited_partner_details: must not exist when company type is limited",
+//                "include_principal_place_of_business_details: must not exist when company type is limited",
+//                "include_general_nature_of_business_information: must not exist when company type is limited"));
         assertThat(errors, containsInAnyOrder(expectedErrors));
     }
 
     @Test
-    @DisplayName("Company objects and good standing set as null when certificate type is dissolution")
+    @DisplayName(
+            "Company objects and good standing set as null when certificate type is dissolution")
     void companyObjectsGoodStandingAsNullWhenRequestedForDissolution() {
         // Given
         certificateItemOptions.setCertificateType(DISSOLUTION);
         certificateItemOptions.setIncludeCompanyObjectsInformation(null);
         certificateItemOptions.setIncludeGoodStandingInformation(null);
         certificateItemOptions.setCompanyType("any");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
 
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
         // Then
-        assertThat(errors, empty());
+        assertThat(errors, is(empty()));
     }
 
     @Test
@@ -215,9 +231,8 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         // Given
         certificateItemOptions.setDeliveryTimescale(SAME_DAY);
         certificateItemOptions.setIncludeEmailCopy(true);
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
         certificateItemOptions.setCompanyType("any");
-	
+
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
@@ -229,13 +244,12 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     @DisplayName("Do not include email copy for standard delivery timescale")
     void doNotIncludeEmailCopyForStandardDeliveryTimescale() {
         // Given
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
         certificateItemOptions.setDeliveryTimescale(STANDARD);
         certificateItemOptions.setIncludeEmailCopy(true);
         certificateItemOptions.setCompanyType("any");
         ApiError expectedError = ApiErrors.raiseError(ApiErrors.ERR_INCLUDE_EMAIL_COPY_NOT_ALLOWED,
                 "include_email_copy: can only be true when delivery timescale is same_day");
-	
+
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
@@ -248,49 +262,53 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     void doNotIncludeOtherDetailsWithoutBasicInformation() {
         // Given
         DirectorOrSecretaryDetails directorOrSecretaryDetails = new DirectorOrSecretaryDetails();
+        directorOrSecretaryDetails.setIncludeAddress(true);
+        directorOrSecretaryDetails.setIncludeAppointmentDate(true);
+        directorOrSecretaryDetails.setIncludeCountryOfResidence(true);
         directorOrSecretaryDetails.setIncludeDobType(PARTIAL);
-        directorOrSecretaryDetails.setIncludeBasicInformation(false);
-
-        DirectorOrSecretaryDetails secretaryDetails = new DirectorOrSecretaryDetails();
-        secretaryDetails.setIncludeAddress(true);
-        secretaryDetails.setIncludeAppointmentDate(true);
-        secretaryDetails.setIncludeCountryOfResidence(true);
-        secretaryDetails.setIncludeDobType(PARTIAL);
-        secretaryDetails.setIncludeNationality(true);
-        secretaryDetails.setIncludeOccupation(true);
-        secretaryDetails.setIncludeBasicInformation(true);
+        directorOrSecretaryDetails.setIncludeNationality(true);
+        directorOrSecretaryDetails.setIncludeOccupation(true);
 
         DesignatedMemberDetails designatedMemberDetails = new DesignatedMemberDetails();
         designatedMemberDetails.setIncludeAddress(true);
         designatedMemberDetails.setIncludeAppointmentDate(true);
         designatedMemberDetails.setIncludeCountryOfResidence(true);
         designatedMemberDetails.setIncludeDobType(PARTIAL);
-        designatedMemberDetails.setIncludeBasicInformation(true);
 
         MemberDetails memberDetails = new MemberDetails();
         memberDetails.setIncludeAddress(true);
         memberDetails.setIncludeAppointmentDate(true);
         memberDetails.setIncludeCountryOfResidence(true);
         memberDetails.setIncludeDobType(PARTIAL);
-        memberDetails.setIncludeBasicInformation(true);
-        
+
         certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
-        certificateItemOptions.setSecretaryDetails(secretaryDetails);
+        certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
         certificateItemOptions.setDesignatedMemberDetails(designatedMemberDetails);
         certificateItemOptions.setMemberDetails(memberDetails);
         certificateItemOptions.setCompanyType("limited");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
-        List<ApiError> expectedErrors = Arrays.asList(
-                ApiErrors.raiseError(ApiErrors.ERR_DESIGNATED_MEMBERS_DETAILS_SUPPLIED, "include_designated_member_details: must not exist when company type is limited"),
-                ApiErrors.raiseError(ApiErrors.ERR_MEMBERS_DETAILS_SUPPLIED, "include_member_details: must not exist when company type is limited"),
-                ApiErrors.raiseError(ApiErrors.ERR_DIRECTOR_DETAILS_SUPPLIED, "director_details: include_dob_type must not be non-null when include_basic_information is false")
-        );
 
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
         // Then
-        assertThat(errors, contains(expectedErrors));
+        assertThat(errors, contains(
+                ApiErrors.raiseError(ApiErrors.ERR_DESIGNATED_MEMBERS_DETAILS_SUPPLIED,"include_designated_member_details: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_MEMBERS_DETAILS_SUPPLIED,"include_member_details: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_DIRECTOR_DETAILS_SUPPLIED,"director_details: include_address, include_appointment_date, include_country_of_residence,"
+                        + " include_nationality, include_occupation must not be true when include_basic_information"
+                        + " is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_DIRECTOR_DETAILS_SUPPLIED,"director_details: include_dob_type must not be non-null when include_basic_information is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_SECRETARY_DETAILS_SUPPLIED,"secretary_details: include_address, include_appointment_date, include_country_of_residence,"
+                        + " include_nationality, include_occupation must not be true when include_basic_information"
+                        + " is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_SECRETARY_DETAILS_SUPPLIED,"secretary_details: include_dob_type must not be non-null when include_basic_information is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_DESIGNATED_MEMBERS_DETAILS_SUPPLIED,"designated_member_details: include_address, include_appointment_date, include_country_of_residence"
+                        + " must not be true when include_basic_information is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_DESIGNATED_MEMBERS_DETAILS_SUPPLIED,"designated_member_details: include_dob_type must not be non-null when "
+                        + "include_basic_information is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_MEMBERS_DETAILS_SUPPLIED,"member_details: include_address, include_appointment_date, include_country_of_residence"
+                        + " must not be true when include_basic_information is false"),
+                ApiErrors.raiseError(ApiErrors.ERR_MEMBERS_DETAILS_SUPPLIED,"member_details: include_dob_type must not be non-null when include_basic_information is false")));
     }
 
     @Test
@@ -305,11 +323,10 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         directorOrSecretaryDetails.setIncludeDobType(PARTIAL);
         directorOrSecretaryDetails.setIncludeNationality(true);
         directorOrSecretaryDetails.setIncludeOccupation(true);
-        
+
         certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
         certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
         certificateItemOptions.setCompanyType("limited");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
 
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
@@ -319,9 +336,12 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     }
 
     @Test
-    @DisplayName("Request is valid if company type is llp and appropriate fields set")
-    void allowMembersAndDesignatedMembersFieldValuesForLps() {
+    @DisplayName("Request is invalid if company type is llp and appropriate fields set")
+    void rejectMembersAndDesignatedMembersFieldValuesForLlps() {
         //given
+        DirectorOrSecretaryDetails directorOrSecretaryDetails = new DirectorOrSecretaryDetails();
+        directorOrSecretaryDetails.setIncludeBasicInformation(true);
+
         MemberDetails memberDetails = new MemberDetails();
         memberDetails.setIncludeAddress(true);
         memberDetails.setIncludeDobType(PARTIAL);
@@ -335,17 +355,22 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         designatedMemberDetails.setIncludeAppointmentDate(true);
         designatedMemberDetails.setIncludeCountryOfResidence(true);
         designatedMemberDetails.setIncludeBasicInformation(true);
-	
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
+
+        certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
+        certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
         certificateItemOptions.setMemberDetails(memberDetails);
         certificateItemOptions.setDesignatedMemberDetails(designatedMemberDetails);
         certificateItemOptions.setCompanyType("llp");
+        List<ApiError> expectedErrors = Arrays.asList(
+                ApiErrors.raiseError(ApiErrors.ERR_DESIGNATED_MEMBERS_DETAILS_SUPPLIED, "include_designated_member_details: must not exist when company type is llp"),
+                ApiErrors.raiseError(ApiErrors.ERR_MEMBERS_DETAILS_SUPPLIED, "include_member_details: must not exist when company type is llp")
+        );
 
         //when
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
         //then
-        assertThat(errors, is(empty()));
+        assertThat(errors, containsInAnyOrder(expectedErrors));
     }
 
     @Test
@@ -361,7 +386,7 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         PrincipalPlaceOfBusinessDetails principalPlaceOfBusinessDetails = new PrincipalPlaceOfBusinessDetails();
         principalPlaceOfBusinessDetails.setIncludeDates(true);
         principalPlaceOfBusinessDetails.setIncludeAddressRecordsType(IncludeAddressRecordsType.CURRENT);
-	
+
         certificateItemOptions.setGeneralPartnerDetails(generalPartnerDetails);
         certificateItemOptions.setLimitedPartnerDetails(limitedPartnerDetails);
         certificateItemOptions.setPrincipalPlaceOfBusinessDetails(principalPlaceOfBusinessDetails);
@@ -384,12 +409,10 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
         directorOrSecretaryDetails.setIncludeAppointmentDate(false);
         directorOrSecretaryDetails.setIncludeNationality(true);
         directorOrSecretaryDetails.setIncludeOccupation(true);
-        directorOrSecretaryDetails.setIncludeBasicInformation(false);
-	
+
         certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
         certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
         certificateItemOptions.setCompanyType("limited");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
         List<ApiError> expectedErrors = Arrays.asList(
                 ApiErrors.raiseError(new ApiError("include-address-error", "director_details.include_address", ApiErrors.BOOLEAN_LOCATION_TYPE, ApiErrors.ERROR_TYPE_VALIDATION), "include_address: must not be true when include_basic_information is false"),
                 ApiErrors.raiseError(new ApiError("include-nationality-error", "director_details.include_nationality", ApiErrors.BOOLEAN_LOCATION_TYPE, ApiErrors.ERROR_TYPE_VALIDATION), "include_nationality: must not be true when include_basic_information is false"),
@@ -410,8 +433,7 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     @DisplayName("Handles absence of item options smoothly")
     void handlesAbsenceOfItemOptionsSmoothly() {
         // Given
-        certificateItemOptions.setCompanyType("limited");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
+        when(requestValidatable.getItemOptions()).thenReturn(null);
 
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
@@ -425,7 +447,6 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     void handlesAbsenceOfDetailsSmoothly() {
         // Given
         certificateItemOptions.setCompanyType("any");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
 
         // When
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
@@ -435,31 +456,25 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
     }
 
     @Test
-    @DisplayName("Request is invalid if directors' or secretaries' details specified for an llp")
-    void rejectDirectorsOrSecretariesDetailsIfSpecifiedForLlp() {
+    @DisplayName("Request is valid if directors' or secretaries' details specified for an llp")
+    void allowDirectorsOrSecretariesDetailsWhenSpecifiedForLlp() {
         //given
         DirectorOrSecretaryDetails directorOrSecretaryDetails = new DirectorOrSecretaryDetails();
 
         certificateItemOptions.setCompanyType("llp");
-        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(null);
         certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
         certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
-        List<ApiError> expectedErrors = Arrays.asList(
-                ApiErrors.raiseError(ApiErrors.ERR_DIRECTOR_DETAILS_SUPPLIED, "include_director_details: must not exist when company type is llp"),
-                ApiErrors.raiseError(ApiErrors.ERR_SECRETARY_DETAILS_SUPPLIED, "include_secretary_details: must not exist when company type is llp")
-        );
 	
         //when
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
         //then
-        assertThat(errors, contains(expectedErrors));
+        assertThat(errors, empty());
     }
 
     @Test
     @DisplayName(
-            "Request is invalid if directors' or secretaries' details specified for an "
-                    + "limited-partnership")
+            "Request is invalid if directors' or secretaries' details specified for an limited-partnership")
     void rejectDirectorsOrSecretariesDetailsIfSpecifiedForLimitedPartnership() {
         //given
         DirectorOrSecretaryDetails directorOrSecretaryDetails = new DirectorOrSecretaryDetails();
@@ -471,11 +486,65 @@ class CreateItemRequestValidatorFeatureFlagsEnabledTest {
                 ApiErrors.raiseError(ApiErrors.ERR_DIRECTOR_DETAILS_SUPPLIED, "include_director_details: must not exist when company type is limited-partnership"),
                 ApiErrors.raiseError(ApiErrors.ERR_SECRETARY_DETAILS_SUPPLIED, "include_secretary_details: must not exist when company type is limited-partnership")
         );
-
-	    //when
+	
+        //when
         final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
 
         //then
         assertThat(errors, contains(expectedErrors));
+    }
+
+    @Test
+    @DisplayName("Request is invalid if LP and LLP details specified for an limited company")
+    void rejectLPAndLLPDetailsIfSpecifiedForLimitedCompany() {
+        //given
+        MemberDetails memberDetails = new MemberDetails();
+        memberDetails.setIncludeAddress(true);
+        memberDetails.setIncludeDobType(PARTIAL);
+        memberDetails.setIncludeAppointmentDate(true);
+        memberDetails.setIncludeCountryOfResidence(true);
+        memberDetails.setIncludeBasicInformation(true);
+
+        DesignatedMemberDetails designatedMemberDetails = new DesignatedMemberDetails();
+        designatedMemberDetails.setIncludeAddress(true);
+        designatedMemberDetails.setIncludeDobType(PARTIAL);
+        designatedMemberDetails.setIncludeAppointmentDate(true);
+        designatedMemberDetails.setIncludeCountryOfResidence(true);
+        designatedMemberDetails.setIncludeBasicInformation(true);
+
+        GeneralPartnerDetails generalPartnerDetails = new GeneralPartnerDetails();
+        generalPartnerDetails.setIncludeBasicInformation(true);
+
+        LimitedPartnerDetails limitedPartnerDetails = new LimitedPartnerDetails();
+        limitedPartnerDetails.setIncludeBasicInformation(true);
+
+        PrincipalPlaceOfBusinessDetails principalPlaceOfBusinessDetails = new PrincipalPlaceOfBusinessDetails();
+        principalPlaceOfBusinessDetails.setIncludeDates(true);
+        principalPlaceOfBusinessDetails.setIncludeAddressRecordsType(IncludeAddressRecordsType.CURRENT);
+
+        DirectorOrSecretaryDetails directorOrSecretaryDetails = new DirectorOrSecretaryDetails();
+
+        certificateItemOptions.setGeneralPartnerDetails(generalPartnerDetails);
+        certificateItemOptions.setLimitedPartnerDetails(limitedPartnerDetails);
+        certificateItemOptions.setPrincipalPlaceOfBusinessDetails(principalPlaceOfBusinessDetails);
+        certificateItemOptions.setIncludeGeneralNatureOfBusinessInformation(true);
+        certificateItemOptions.setMemberDetails(memberDetails);
+        certificateItemOptions.setDesignatedMemberDetails(designatedMemberDetails);
+        certificateItemOptions.setCompanyType("limited");
+        certificateItemOptions.setDirectorDetails(directorOrSecretaryDetails);
+        certificateItemOptions.setSecretaryDetails(directorOrSecretaryDetails);
+
+        //when
+        final List<ApiError> errors = validatorUnderTest.getValidationErrors(requestValidatable);
+
+        //then
+        assertThat(errors, containsInAnyOrder(
+                ApiErrors.raiseError(ApiErrors.ERR_PRINCIPAL_PLACE_OF_BUSINESS_DETAILS_SUPPLIED, "include_principal_place_of_business_details: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_GENERAL_PARTNER_DETAILS_SUPPLIED, "include_general_partner_details: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_LIMITED_PARTNER_DETAILS_SUPPLIED, "include_limited_partner_details: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_INCLUDE_GENERAL_NATURE_OF_BUSINESS_INFORMATION_SUPPLIED, "include_general_nature_of_business_information: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_DESIGNATED_MEMBERS_DETAILS_SUPPLIED, "include_designated_member_details: must not exist when company type is limited"),
+                ApiErrors.raiseError(ApiErrors.ERR_MEMBERS_DETAILS_SUPPLIED, "include_member_details: must not exist when company type is limited")
+        ));
     }
 }
